AD8307 add-on
In the settings.h can be selected if the AD8307 board is used for the power measurements. 
In the "Services" menu is added an option to control the scaling of the power meter circuit.
When the AD8307 is selected, are appearing the formulas how to calibrate the measurement. Once calculated the values need to be saved in the calibration menu. 
The procedure is repeated for the FW and the reflected power measurement
Then you just enjoy the accurate measurements :)